Refactor MotionZoom and MotionTranslate speeds to be floating-point

+++ libbrot2/MovieMotion.cpp
@@ -56,19 +56,19 @@ struct signpair calc_signs(const Fractal::Point a, const Fractal::Point b)
  * we move the corners by up to @speed@ pixels.
  * Returns true if we did something, false if not. */
 bool Movie::MotionZoom(const Fractal::Point& size_in, const Fractal::Point& size_target,
-		const unsigned width, const unsigned height, const unsigned speed, Fractal::Point& size_out)
+		const unsigned width, const unsigned height, const float speed, Fractal::Point& size_out)
 {
 	// Easy case
 	size_out = size_in;
 	if ( real(size_in) == real(size_target)  ||  imag(size_in) == imag(size_target) )
 		return false;
 
-	int speed_z = speed * width / 300;
+	float speed_z = speed * width / 300;
 	// If we are zooming IN, size is DECREASING, so the delta is NEGATIVE 
 	if (real (size_in) > real(size_target) )
 		speed_z = -speed_z;
-	int speed_x = speed_z;
-	double speed_y = (double) speed_x * height / width;
+	float speed_x = speed_z;
+	float speed_y = (float) speed_x * height / width;
 
 	// If we're close enough to target, we're done.
 	// But what does "close enough" mean? Good old floating-point comparisons...
@@ -94,7 +94,7 @@ bool Movie::MotionZoom(const Fractal::Point& size_in, const Fractal::Point& size
  * Returns true if we did something, false if not. */
 bool Movie::MotionTranslate(const Fractal::Point& centre_in, const Fractal::Point& centre_target,
 		const Fractal::Point& size,
-		const unsigned width, const unsigned height, const unsigned speed, Fractal::Point& centre_out)
+		const unsigned width, const unsigned height, const float speed, Fractal::Point& centre_out)
 {
 
 	centre_out = centre_in;
@@ -106,8 +106,8 @@ bool Movie::MotionTranslate(const Fractal::Point& centre_in, const Fractal::Poin
 	struct signpair signs_before(calc_signs(centre_target, centre_in));
 	Fractal::Point tmp_out;
 
-	unsigned speed_x = speed * width / 300;
-	double speed_y = (double) speed_x * height / width;
+	float speed_x = speed * width / 300;
+	float speed_y = (float) speed_x * height / width;
 	Fractal::Point pixel_size ( real(size) / width, imag(size) / height );
 	Fractal::Point delta ( real(pixel_size) * speed_x, imag(pixel_size) * speed_y );
